# Optimize dataset detail view for dataset with many resources